refactor(1단계 리팩터링): id를 Int에서 Long으로 변경

+++ domain/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -5,10 +5,10 @@ import woowacourse.shopping.domain.util.WoowaResult
 
 interface ProductRepository {
 
-    fun fetchProduct(callback: (WoowaResult<CartProduct>) -> Unit, id: Int)
+    fun fetchProduct(callback: (WoowaResult<CartProduct>) -> Unit, id: Long)
     fun fetchPagedProducts(
         callback: (products: WoowaResult<List<CartProduct>>, isLast: Boolean) -> Unit,
         pageItemCount: Int,
-        lastId: Int,
+        lastId: Long,
     )
 }
